% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmBinomial.R
\name{glmBinomial}
\alias{glmBinomial}
\title{Perform diagnostics and obtain results for a fitted binomial regression model.}
\usage{
glmBinomial(myObject)
}
\arguments{
\item{myObject}{Output of glmSetup function}
}
\value{
A list containing various diagnostic results and information about the fitted binomial regression model.
}
\description{
This function takes the output from glmSetup() and performs diagnostics for a binomial regression model.
}
\examples{
# Example usage:
glm_full <- glmSetup(res_var = "Count", base_var = "TNF", input_data = Cell_Differentiation, tolerance = 0.00001, log_offset = 0.1, max_iterations = 25)
Binomial_diagnostics <- glmBinomial(glm_full)

}
