% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPoisson.R
\name{glmPoisson}
\alias{glmPoisson}
\title{Perform diagnostics and obtain results for a fitted Poisson regression model.}
\usage{
glmPoisson(myObject)
}
\arguments{
\item{myObject}{Output of glmSetup function}
}
\value{
A list containing various diagnostic results and information about the fitted Poisson regression model.
}
\description{
This function takes the output from glmSetup() and performs diagnostics for a Poisson regression model.
}
\examples{
# Example usage:
Cell_Differentiation <- read.csv(system.file("data", "Cell_Differentiation.csv", package = "redo.glm"))
glm_full <- glmSetup(res_var = "Count", base_var = "TNF", input_data = Cell_Differentiation, tolerance = 0.00001, log_offset = 0.1, max_iterations = 25)
glmPoisson(glm_full)

}
