% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSetup.R
\name{glmSetup}
\alias{glmSetup}
\title{Set up for IRWLS generalized linear model}
\usage{
glmSetup(res_var, base_var, input_data, tolerance, log_offset, max_iterations)
}
\arguments{
\item{res_var}{Response variable name}

\item{base_var}{Baseline variable name}

\item{input_data}{Input data frame}

\item{tolerance}{Tolerance for convergence}

\item{log_offset}{Offset for log transformation}

\item{max_iterations}{Maximum number of iterations}
}
\value{
List containing setup information
}
\description{
This function prepares the data and parameters for fitting a generalized
linear model using Iteratively Reweighted Least Squares (IRWLS). Note that
this model setup currently does not handle interaction terms.
}
\examples{
# Example usage:
glm_full <- glmSetup(res_var = "Count", base_var = "TNF", input_data = Cell_Differentiation, tolerance = 0.00001, log_offset = 0.1, max_iterations = 25)

}
\seealso{
Use glmPoisson() or glmBinomial() to perform model diagnostics and obtain detailed results.
}
